almost finish chunk enc/dec
finish AEAD udp handling
need more thinking about TCP handing
no error checks now

+++ fuckshadows-csharp/Encryption/Sodium.cs
@@ -67,16 +67,16 @@ namespace Fuckshadows.Encryption
             ulong mlen, IntPtr ad, ulong adlen, IntPtr nsec, byte[] npub, byte[] k);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int crypto_aead_chacha20poly1305_ietf_decrypt(byte[] m, ref int mlen_p,
-            IntPtr nsec, byte[] c, int clen, IntPtr ad, int adlen, byte[] npub, byte[] k);
+        public static extern int crypto_aead_chacha20poly1305_ietf_decrypt(byte[] m, ref ulong mlen_p,
+            IntPtr nsec, byte[] c, ulong clen, IntPtr ad, ulong adlen, byte[] npub, byte[] k);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
         public static extern int crypto_aead_chacha20poly1305_encrypt(byte[] c, ref ulong clen_p, byte[] m, ulong mlen,
             IntPtr ad, ulong adlen, IntPtr nsec, byte[] npub, byte[] k);
 
         [DllImport(DLLNAME, CallingConvention = CallingConvention.Cdecl)]
-        public static extern int crypto_aead_chacha20poly1305_decrypt(byte[] m, ref int mlen_p, IntPtr nsec, byte[] c,
-            int clen, IntPtr ad, int adlen, byte[] npub, byte[] k);
+        public static extern int crypto_aead_chacha20poly1305_decrypt(byte[] m, ref ulong mlen_p, IntPtr nsec, byte[] c,
+            ulong clen, IntPtr ad, ulong adlen, byte[] npub, byte[] k);
 
         #endregion
